swapped who's wrapped and who's wrapping
